+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - add JWT auth for the `cms-api`
  - add `refresh_token` functionality for JWT login for better security
  - remove `open`page type, and all pages from that type were moved to `experiment` with `open_access` flag
+ - add column `is_system`, where if enabled the page is system and it should not be delete it or change its properties. Only its content fields can be edited
 
 ### Bugfix
  - fix the way that the data is decoded in the `formUserInput` controller